+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## [1.139.4](https://github.com/carlonicora/nextjs-jsonapi/compare/v1.139.3...v1.139.4) (2026-07-27)
+
+### 🐛 Bug Fixes
+
+* **components:** declare non-native button on EditorSheet trigger ([47d4aad](https://github.com/carlonicora/nextjs-jsonapi/commit/47d4aad9a1995fed6f1a8d0893d1c34598f6f2ee))
+
+## [1.139.3](https://github.com/carlonicora/nextjs-jsonapi/compare/v1.139.2...v1.139.3) (2026-07-27)
+
+### 🐛 Bug Fixes
+
+* **navigations:** keep RecentPagesNavigator labelled in the mobile sheet ([5273499](https://github.com/carlonicora/nextjs-jsonapi/commit/527349968d6305a35586a2cd95938db43f39756d))
+
+## [1.139.2](https://github.com/carlonicora/nextjs-jsonapi/compare/v1.139.1...v1.139.2) (2026-07-27)
+
+### 🐛 Bug Fixes
+
+* **oauth:** make scope optional in the consent flow per RFC 6749 §3.3 ([7e5be8d](https://github.com/carlonicora/nextjs-jsonapi/commit/7e5be8d54b8ff508c9c47106f889fbf10f792569))
+
 ## [1.139.1](https://github.com/carlonicora/nextjs-jsonapi/compare/v1.139.0...v1.139.1) (2026-07-25)
 
 ### 🐛 Bug Fixes